rename(friend): ui 레이어 model 패키지 분리
